+++ packages/atomic/cypress/integration/facets/category-facet.cypress.ts
@@ -544,7 +544,6 @@ describe('When URL contains a selected path of category facet', () => {
       .find('ul[part="parents"]')
       .find('li')
       .last()
-      .find('b')
       .contains('Alberta');
   });
 });
@@ -619,5 +618,21 @@ describe('Category Facet with custom basePath and custom filterByBasePath', () =
     cy.getTextOfAllElements('@StandardFacetAllValues').then((labels) => {
       expect(labels.length).to.greaterThan(1);
     });
+  });
+});
+
+describe('when no first search has yet been executed', () => {
+  beforeEach(() => {
+    setUpPage(
+      ` <atomic-category-facet field="${categoryFacetProp.field}"></atomic-category-facet>`,
+      false
+    );
+  });
+
+  it('should render a placeholder', () => {
+    cy.get(FacetSelectors.categoryFacet)
+      .shadow()
+      .find('div[part="placeholder"]')
+      .should('be.visible');
   });
 });