+++ packages/atomic/src/components.d.ts
@@ -73,6 +73,10 @@ export namespace Components {
           * The non-localized label for the facet
          */
         "label": string;
+        /**
+          * The number of values to request for this facet, when there are no manual ranges.
+         */
+        "numberOfValues": number;
     }
     interface AtomicDateRange {
         /**
@@ -145,6 +149,10 @@ export namespace Components {
           * The non-localized label for the facet.
          */
         "label": string;
+        /**
+          * The number of values to request for this facet, when there are no manual ranges.
+         */
+        "numberOfValues": number;
     }
     interface AtomicNumericRange {
         /**
@@ -701,6 +709,10 @@ declare namespace LocalJSX {
           * The non-localized label for the facet
          */
         "label"?: string;
+        /**
+          * The number of values to request for this facet, when there are no manual ranges.
+         */
+        "numberOfValues"?: number;
     }
     interface AtomicDateRange {
         /**
@@ -773,6 +785,10 @@ declare namespace LocalJSX {
           * The non-localized label for the facet.
          */
         "label"?: string;
+        /**
+          * The number of values to request for this facet, when there are no manual ranges.
+         */
+        "numberOfValues"?: number;
     }
     interface AtomicNumericRange {
         /**