+++ packages/atomic/src/components/facets/base-facet/base-facet.tsx
@@ -56,14 +56,11 @@ export const BaseFacet: FunctionalComponent<BaseFacetProps> = (
     </button>
   ) : null;
 
-  const fullWidthEllipsedClasses =
-    'overflow-ellipsis whitespace-nowrap overflow-hidden w-full';
-
   return (
     <div class="facet mb-4" part="facet">
       <button
         title={props.label}
-        class={`facet-button text-left border-solid bg-background px-4 h-9 outline-none focus:outline-none lg:hidden cursor-pointer ${fullWidthEllipsedClasses} ${
+        class={`facet-button text-left border-solid bg-background px-4 h-9 outline-none focus:outline-none lg:hidden cursor-pointer ellipsed w-full ${
           props.hasActiveValues
             ? 'border-2 border-primary text-primary'
             : 'border border-divider text-on-background-variant'
@@ -80,7 +77,7 @@ export const BaseFacet: FunctionalComponent<BaseFacetProps> = (
         <div class="flex flex-row items-center pb-2 mb-2 border-b border-solid border-divider">
           <span
             title={props.label}
-            class={`font-semibold text-on-background-variant text-base lg:text-sm ${fullWidthEllipsedClasses}`}
+            class="font-semibold text-on-background-variant text-base lg:text-sm ellipsed w-full"
           >
             {props.label}
           </span>