+++ packages/headless/src/controllers/facets/facet/headless-facet.ts
@@ -42,6 +42,13 @@ import {
 } from './headless-facet-options';
 import {determineFacetId} from '../_common/facet-id-determinor';
 import {FacetValueState} from '../../../features/facets/facet-api/value';
+import {
+  configuration,
+  facetSearchSet,
+  facetSet,
+  search,
+} from '../../../app/reducers';
+import {loadReducerError} from '../../../utils/errors';
 
 export {FacetOptions, FacetSearchOptions, FacetValueState};
 
@@ -226,12 +233,11 @@ export interface FacetValue {
  * @param props - The configurable `Facet` properties.
  * @returns A `Facet` controller instance.
  * */
-export function buildFacet(
-  engine: Engine<
-    FacetSection & ConfigurationSection & FacetSearchSection & SearchSection
-  >,
-  props: FacetProps
-): Facet {
+export function buildFacet(engine: Engine<unknown>, props: FacetProps): Facet {
+  if (!loadFacetReducers(engine)) {
+    throw loadReducerError;
+  }
+
   const {dispatch} = engine;
   const controller = buildController(engine);
 
@@ -267,6 +273,8 @@ export function buildFacet(
 
     return initialNumberOfValues < currentValues.length && hasIdleValues;
   };
+
+  const getIsLoading = () => engine.state.search.isLoading;
 
   dispatch(registerFacet(options));
   const facetSearch = createFacetSearch();
@@ -330,7 +338,7 @@ export function buildFacet(
       const request = getRequest();
       const response = getResponse();
 
-      const isLoading = engine.state.search.isLoading;
+      const isLoading = getIsLoading();
       const sortCriterion = request.sortCriteria;
       const values = response ? response.values : [];
       const hasActiveValues = values.some(
@@ -350,4 +358,13 @@ export function buildFacet(
       };
     },
   };
+}
+
+function loadFacetReducers(
+  engine: Engine<unknown>
+): engine is Engine<
+  FacetSection & ConfigurationSection & FacetSearchSection & SearchSection
+> {
+  engine.addReducers({facetSet, configuration, facetSearchSet, search});
+  return true;
 }