+++ packages/headless/src/features/facets/facet-search-set/facet-search-reducer-helpers.test.ts
@@ -35,7 +35,7 @@ describe('FacetSearch slice', () => {
         facetId,
         captions: {},
         numberOfValues: 10,
-        query: '',
+        query: '**',
       });
     });